==> 01_Create_Dataset.ipynb <==
# Installation

Use this: https://riverml.xyz/latest/api/tree/iSOUPTreeRegressor/

# Install a pip package in the current Jupyter kernel
# import sys
# !{sys.executable} -m pip install river --user

import numpy
import pandas as pd
import json
from datetime import datetime

# load static data
def load_static():
    static_df = pd.read_csv(".\\data\\static\\staticFeatures.csv", sep=';', encoding='utf-8', index_col=None)
    static_df.drop('Unnamed: 10', axis='columns', inplace=True)
    static_df['timestamp'] = pd.to_datetime(static_df['timestamp'], infer_datetime_format=True)
    static_df.set_index('timestamp', inplace=True)
    return static_df

static_df = load_static()

# flatten json
def flatten_json(y):
    out = {}
    
    def flatten(x, name=''):
        if type(x) is dict:
            for a in x:
                flatten(x[a], a + name)
        elif type(x) is list:
            i = 0
            for a in x:
                flatten(a, name + str(i) + "_")
                i += 1
        else:
            out[name[:-1]] = x
            
    flatten(y)
    return out

# load weather data
def load_weather():
    with open(".\\data\\weather\\converted.json") as f:
        lines = f.read().splitlines()

    df_inter = pd.DataFrame(lines)
    df_inter.columns = ['json_element']
    df_inter['json_element'].apply(json.loads)
    df_final = pd.json_normalize(df_inter['json_element'].apply(json.loads))
    for j in range(df_final.shape[0]):
        print("\r" + str(j), end="")
        for i in range(49):
            if i != 0:
                df_final['hourly.data'][j][i].pop("time", None)
            df_final['hourly.data'][j][i].pop("icon", None)
            df_final['hourly.data'][j][i].pop("apparentTemperature", None)
            df_final['hourly.data'][j][i].pop("uvIndex", None)
            df_final['hourly.data'][j][i].pop("precipType", None)
            df_final['hourly.data'][j][i].pop("summary", None)
            df_final['hourly.data'][j][i].pop("precipIntensity", None)
            df_final['hourly.data'][j][i].pop("precipProbability", None)
            """
            df_final['hourly.data'][0][i]["temperature" + str(i)] = df_final['hourly.data'][0][i].pop("temperature")
            df_final['hourly.data'][0][i]["dewPoint" + str(i)] = df_final['hourly.data'][0][i].pop("dewPoint")
            df_final['hourly.data'][0][i]["humidity" + str(i)] = df_final['hourly.data'][0][i].pop("humidity")
            df_final['hourly.data'][0][i]["pressure" + str(i)] = df_final['hourly.data'][0][i].pop("pressure")
            df_final['hourly.data'][0][i]["windSpeed" + str(i)] = df_final['hourly.data'][0][i].pop("windSpeed")
            df_final['hourly.data'][0][i]["windBearing" + str(i)] = df_final['hourly.data'][0][i].pop("windBearing")
            df_final['hourly.data'][0][i]["cloudCover" + str(i)] = df_final['hourly.data'][0][i].pop("cloudCover", 0.0)
            df_final['hourly.data'][0][i]["visibility" + str(i)] = df_final['hourly.data'][0][i].pop("visibility")
            """

        df_final['hourly.data'][j] = flatten_json(df_final['hourly.data'][j])

    weather_df = pd.json_normalize(df_final['hourly.data'])

    # remove ozone and windGust
    columns = list(weather_df.columns)
    for a in columns:
        if a.find("ozone") != -1:
            weather_df.pop(a)
        if a.find("windGust") != -1:
            weather_df.pop(a)

    # fill in missing values
    weather_df.interpolate(axis=0, limit=None, inplace=True)
    weather_df.fillna(method="backfill", inplace=True)
    weather_df.fillna(method="ffill", inplace=True)

    weather_df['timestamp'] = pd.to_datetime(weather_df['time0']*1000*1000*1000)
    foo = weather_df.pop('time0')
    weather_df.set_index('timestamp', inplace=True)

    # remove duplicates
    weather_df = weather_df.loc[~weather_df.index.duplicated(keep='first')]
    
    return weather_df

weather_df = load_weather()

def load_sensor(n):
    sensor_df = pd.read_json('./data/smartmeters/N' + str(n) + '.json')
    sensor_df.pop("_id")
    sensor_df.pop("node_id")
    sensor_df.pop("stamp_db")
    sensor_df.pop("pg")
    sensor_df.pop("qc")
    sensor_df.pop("qg")
    sensor_df['timestamp'] = pd.to_datetime(sensor_df['stamp'] * 1000 * 1000 * 1000) # unix ts in nanoseconds
    foo = sensor_df.pop("stamp")
    sensor_df.set_index('timestamp', inplace=True)
    
    # enrich sensor_df
    # add 15, 30, 45-minute values
    sensor_df["pc15"] = sensor_df.shift(periods=-1, freq='0.25H')["pc"]
    sensor_df["pc30"] = sensor_df.shift(periods=-2, freq='0.25H')["pc"]
    sensor_df["pc45"] = sensor_df.shift(periods=-3, freq='0.25H')["pc"]
    sensor_df = sensor_df[:-3]

    # add historic values
    moving_averages = ['1H', '6H', '1D', '7D', '30D']
    delays = ['1H', '2H', '3H', '6H', '12H', '1D', '2D', '3D', '7D']

    # create moving averages
    for a in moving_averages:    
        sensor_df['pc_ma_' + a] = sensor_df['pc'].rolling(window=a, min_periods=1).mean()
        sensor_df['pc_std_' + a] = sensor_df['pc'].rolling(window=a, min_periods=1).std()

        # historic values
    for a in moving_averages: 
        for d in delays:
            sensor_df["pc_ma_{}_{}".format(a, d)] = sensor_df.shift(periods=1, freq=d)["pc_ma_{}".format(a)]
    
    return sensor_df

def make_final(n):
    sensor_df = load_sensor(n)
    final_df = pd.concat([sensor_df, static_df], axis=1, join="inner")
    final2_df = pd.concat([weather_df, final_df], axis=1, join="inner")    
    return final_df.dropna()

make_final(1)

weather_df
