==> 01_Create_DataSet_with_Util_File.ipynb <==
# importing helper functions
from data import *

# loading static data
static_df = load_static()

"""
# loading weather data - this takes a few minutes
weather_df = load_weather()
# save weather
weather_df.to_pickle("data/weather/weather.pkl")
"""

# load weather
weather_df = pd.read_pickle("data/weather/weather.pkl")

# loading and extending making sensor data
final_df = make_final(static_df, weather_df, 1)

weather_df

final_df

list(final_df.columns)
