==> 02_Create_Targets_with_Basic_Model.ipynb <==
# Load Data

from data import *

static_df = load_static()
weather_df = pd.read_pickle('data/weather/weather.pkl')

final_df = make_final(static_df, weather_df, 1)

final_df.isna().sum().sum()

def make_target(df, h):
    y = df.shift(periods=-h, freq='1H')["pc"]
    y = y[h:]
    X = df[:-h]
    return X.to_numpy(), y.to_numpy()

X, y = make_target(final_df, 1)

# Create Simple Model

from sklearn.model_selection import TimeSeriesSplit
from sklearn import linear_model
from sklearn import tree
from sklearn.ensemble import RandomForestRegressor

tscv = TimeSeriesSplit(n_splits = 10)

reg = linear_model.LinearRegression()
#reg = tree.DecisionTreeRegressor()
#reg = RandomForestRegressor(max_depth=20)
i = 0
scores = []
for train, test in tscv.split(X, y):    
    i = i + 1
    print(i)
    reg.fit(X[train], y[train])
    scores.append(reg.score(X[test], y[test]))
scores

rf_scores = scores
rf_scores
