==> 06_Experiments_ScikitLearn_General.ipynb <==
# importing helper functions
from data import *

# import preprocessing tools
from sklearn import preprocessing
import numpy as np
import pandas as pd

# import learning/evaluation
from sklearn.tree import DecisionTreeRegressor
from sklearn.linear_model import LinearRegression
from sklearn.ensemble import RandomForestRegressor
from sklearn.ensemble import GradientBoostingRegressor
from sklearn.neural_network import MLPRegressor

from sklearn.metrics import f1_score
from sklearn.metrics import r2_score
from sklearn.metrics import mean_absolute_percentage_error
from sklearn.model_selection import cross_val_score
from sklearn.model_selection import TimeSeriesSplit

# dataset creation
def make_target(df, h):
    if h == 1024:
        hh = 24
        y = df.shift(periods=-hh, freq='1H')["pc_ma_1D"]
        df.pop('pc15')
        df.pop('pc30')
        df.pop('pc45')
        y = y[hh:]
        X = df[:-hh]
    elif h > 100:
        hh = h - 100
        y = df.shift(periods=-hh, freq='1H')["pc_ma_6H"]
        df.pop('pc15')
        df.pop('pc30')
        df.pop('pc45')
        y = y[hh:]
        X = df[:-hh]
    elif h >= 1:
        y = df.shift(periods=-h, freq='1H')["pc"]
        df.pop('pc15')
        df.pop('pc30')
        df.pop('pc45')
        y = y[h:]
        X = df[:-h]
    else: 
        y = {
            0.25: df["pc15"],
            0.5: df["pc30"],
            0.75: df["pc45"]
        }[h]
        df.pop('pc15')
        df.pop('pc30')
        df.pop('pc45')        
        X = df
    return X, y

# load data
static_df = load_static()
weather_df = pd.read_pickle('data/weather/weather.pkl')

sensors = [1, 24, 35, 43, 68, 103, 105, 106, 108, 109, 110, 116, 117, 122, 124, 144, 145, 157, 168, 177,
    179, 223, 225, 236, 239, 245, 249, 253, 257, 259, 262, 285, 286, 290, 293, 301, 303, 312, 358, 359, 362,
    365, 367, 377, 378, 379, 382, 384, 385, 388, 389, 391, 392, 394, 395]
# horizons = [0.25, 0.5, 0.75, 1, 3, 6, 12, 24, 36]
horizons = [106, 112, 118, 124, 1024]

models = [
    ('LinReg', 'LinearRegression()'),
    ('DT', 'DecisionTreeRegressor()'),
    ('RF', 'RandomForestRegressor(n_jobs=10)'),
    ('GBT', 'GradientBoostingRegressor()')
    #('MLP', 'MLPRegressor(hidden_layer_sizes=(10,))')
]

result = []
for s in sensors:
    for h in horizons:
        for name, model in models: 
            output = "s{}h{}-{}".format(s, h, name)
            print(output)

            # load sensor - last parameter is id
            final_df = make_final(static_df, weather_df, s)

            # make dataset
            X_df, y_df = make_target(final_df, h)
            
            hh = h
            if (hh < 1):
                hh = 1
            selected_features = get_all_candidates(s)
            X = X_df[selected_features].values.astype(float)
            y = y_df.values.astype(float)

            tscv = TimeSeriesSplit(n_splits=50)

            reg = eval(model)

            i = 0            
            partial_result = { "sensor": s, "horizon": h, "model": name, "mape": [], "r2": []}
            for train_index, test_index in tscv.split(X):
                # print("TRAIN:", train_index, "TEST:", test_index)
                i = i + 1    
                X_train, X_test = X[train_index], X[test_index]
                y_train, y_test = y[train_index], y[test_index]

                reg.fit(X_train, y_train)
                y_pred = reg.predict(X_test)

                partial_result["mape"].append(mean_absolute_percentage_error(y_test, y_pred))
                partial_result["r2"].append(r2_score(y_test, y_pred))
                #print(i, r2_score(y_test, y_pred))

            result.append(partial_result)



# save features
with open("results/sklearn_general_additional.pkl", "wb") as f:
    pickle.dump(result, f)

list(final_df.columns)
